feat(plugins): Move Kotlin imports and module_helpers to plugin (JSON)
feat(plugins): Move the rest of Kotlin Bincode to plugin
