+++ crates/facet_generate/tests/swift_generation.rs
@@ -238,7 +238,9 @@ fn set_of_tuple_errors() {
     };
 
     let registry = reflect!(MyStruct).unwrap();
-    let lang = Swift::new(Encoding::Bincode, vec![], &registry);
+    let config = facet_generate::generation::CodeGeneratorConfig::new(String::new())
+        .with_encoding(Encoding::Bincode);
+    let lang = Swift::new(&config, &registry);
     let mut out = Vec::new();
     let mut w = IndentedWriter::new(&mut out, IndentConfig::Space(4));
     let result: std::io::Result<()> = (|| {
@@ -270,7 +272,9 @@ fn map_with_tuple_key_errors() {
     };
 
     let registry = reflect!(MyStruct).unwrap();
-    let lang = Swift::new(Encoding::Bincode, vec![], &registry);
+    let config = facet_generate::generation::CodeGeneratorConfig::new(String::new())
+        .with_encoding(Encoding::Bincode);
+    let lang = Swift::new(&config, &registry);
     let mut out = Vec::new();
     let mut w = IndentedWriter::new(&mut out, IndentConfig::Space(4));
     let result: std::io::Result<()> = (|| {
